2004-04-08  Russell M. Taylor II  <taylorr@cs.unc.edu>
        * Makefile : Adding flag to SGI compile in the hope that it would
                fix things and let us use the new streams libraries.  It
                got us part of the way there on my system.
        * vrpn_BaseClass.C : Added Patrick Hartling's changes to make VRPN
                compile with Visual Studio.NET, which required switching to
                new streams, which required retrofitting SGIs to use old
                streams (using yet another define in vrpn_Configure.h).
        * vrpn_BaseClass.h : Same.
        * vrpn_Clock.C : Same.
        * vrpn_Configure.h : Same.
        * vrpn_Connection.C : Same.
        * vrpn_ImmersionBox. : Same.
        * vrpn_Shared.C : Same.
        * vrpn_TempImager.h : Same.
        * vrpn_Tng3.C : Same.
        * vrpn_Tracker_AnalogFly : Adding Patrick Hartlings changes to make
                it compile with Visual Studio.NET, in this case a type cast
                to let the compiler know which function to call.
        * vrpn_Tracker_ButtonFly : Same.

+++ vrpn_Tracker_AnalogFly.C
@@ -157,9 +157,9 @@ void	vrpn_Tracker_AnalogFly::handle_analog_update
 
 	// Scale and apply the power to the value (maintaining its sign)
 	if (value_offset >=0) {
-		full->value = pow(value_offset*full->axis.scale, full->axis.power);
+		full->value = pow(value_offset*full->axis.scale, (double) full->axis.power);
 	} else {
-		full->value = -pow(value_abs*full->axis.scale, full->axis.power);
+		full->value = -pow(value_abs*full->axis.scale, (double) full->axis.power);
 	}
 }
